2009-05-27 02:06 UTC+0200 Viktor Szakats (harbour.01 syenar hu)   * contrib/hbct/print.c     + added PRINTSEND() support for Windows platform.       (Borrowed from xhb, work of Pavel Tsarenko, with few minor changes)
  * contrib/gtwvg/wvgwin.c
  * contrib/gtwvg/wvgsink.c
    ! Fixed two new warnings after my recent commits.

  * contrib/hbwin/hbwapi.h
    * Tweaked previous commit. COLORREF and HRESULT now using
      hb_retnint() again to fix 32-bit numbers appearing as signed ones
      at .prg level.

  * source/rdd/usrrdd/rdds/arrayrdd.prg
    ; Added some patches recently committed to xhb:
    ! few fixes provided by Quique
    + Added AR_LOCK() function, but commented.
    ; NOTE: Francesco, would you mind reviewing these changes?

  * utils/hbmk2/hbmk2.prg
    * Minor formatting, comment added.

+++ harbour/contrib/gtwvg/wvgsink.c
@@ -383,7 +383,7 @@ hb_ToOutDebug( "event = %i",(int)dispid );
 
          iArg = params->cArgs;
 
-         if( iArg > HB_SIZEOFARRAY( pItemArray ) )
+         if( iArg > ( int ) HB_SIZEOFARRAY( pItemArray ) )
             iArg = HB_SIZEOFARRAY( pItemArray );
 
          for( i = 1; i <= iArg; i++ )
